Move StridedSlice outside experimental
