+++ src/ngraph/op/strided_slice.hpp
